+++ apps/svelte-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # svelte-app
 
+## 0.0.3-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @solar-icons/svelte@1.1.0-beta.1
+
 ## 0.0.3-beta.0
 
 ### Patch Changes