+++ apps/vue-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # vue-app
 
+## 0.0.1-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @solar-icons/vue@1.2.0-beta.1
+
 ## 0.0.1-beta.0
 
 ### Patch Changes