+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solar-icons/nuxt
 
+## 1.2.0-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @solar-icons/vue@1.2.0-beta.1
+
 ## 1.2.0-beta.0
 
 ### Minor Changes